+++ vm/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.2] - 2025-12-13
+
 ## Added
 
 - Add `remove_3rd_party` api
@@ -73,7 +75,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3405]: https://github.com/dusk-network/rusk/issues/3405
 [#3437]: https://github.com/dusk-network/rusk/issues/3437
 
-[Unreleased]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.4.1...HEAD
+[Unreleased]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.4.2...HEAD
+[1.4.2]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.4.1...dusk-vm-1.4.2
 [1.4.1]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.4.0...dusk-vm-1.4.1
 [1.4.0]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.3.0...dusk-vm-1.4.0
 [1.3.0]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.2.0...dusk-vm-1.3.0